+++ flutter_inappwebview_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.10
+
+- Updated `flutter_inappwebview_platform_interface` version dependency to `^1.0.8`
+- Implemented `PlatformCustomPathHandler` class
+
 ## 1.0.9
 
 - Updated `flutter_inappwebview_platform_interface` version dependency to `^1.0.7`