+++ flutter_inappwebview_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.6
+
+- Updated `flutter_inappwebview_platform_interface` version dependency to `^1.0.8`
+
 ## 1.0.5
 
 - Updated `flutter_inappwebview_platform_interface` version dependency to `^1.0.7`